LayoutTests/http/tests/security/contentSecurityPolicy/shared-worker-connect-src-blocked.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="X-WebKit-CSP" content="connect-src 'none'"/>
<script>
if (window.layoutTestController) {
    layoutTestController.waitUntilDone();
    layoutTestController.dumpAsText();
}
</script>
</head>
<body>
<script>
try {
    var worker = new SharedWorker('http://127.0.0.1:8000/security/contentSecurityPolicy/resources/shared-worker-make-xhr.js');
    worker.port.onmessage = function (event) {
        alert(event.data);
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    };
} catch (e) {
    alert(e);
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}
</script>
</body>
</html>
